requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
google-generativeai==0.3.2
yfinance==0.2.35
jinja2==3.1.3
pydantic==2.5.3
python-multipart==0.0.6